move SSRF code to sub-package of util
