+++ util/ssrf/ssrf_test.go
@@ -1,4 +1,4 @@
-package util
+package ssrf
 
 import (
 	"net/http"
